refactor: Message 관련 컴포넌트 리팩토링 (#50)
* refactor: Message.tsx 내부 컴포넌트들을 Message 폴더 내부로 분리

* refactor: 컴포넌트명의 Messages 를 MessageGroup 으로 변경 및 props type 컴포넌트 내부로 이동

* refactor: 단일 Message 관련 컴포넌트들을 폴더로 묶기

* chore: stompjs 추가

* refactor: class변수들을 컴포넌트 외부로 이동
